FRLG Bag Port

+++ include/item_menu_icons.h
@@ -16,7 +16,7 @@ void AddBagItemIconSprite(u16 itemId, u8 id);
 void RemoveBagItemIconSprite(u8 id);
 void CreateItemMenuSwapLine(void);
 void SetItemMenuSwapLineInvisibility(bool8 invisible);
-void UpdateItemMenuSwapLinePos(u8 y);
+void UpdateItemMenuSwapLinePos(u16 y);
 u8 CreateBerryTagSprite(u8 id, s16 x, s16 y);
 void FreeBerryTagSpritePalette(void);
 u8 CreateSpinningBerrySprite(u8 berryId, u8 x, u8 y, bool8 startAffine);
